==> 2_NB_models.ipynb <==
# Naive Bayes with Structure Data

import numpy as np
import pandas as pd
from sklearn.naive_bayes import MultinomialNB
from sklearn.model_selection import train_test_split, cross_validate
from sklearn.metrics import roc_auc_score

StructuralInfo=pd.read_csv("email_05_structure.csv",index_col=[0])
StructuralInfo.describe()

# Shuffle=True
x_train,x_test,y_train,y_test=train_test_split(StructuralInfo[["multipart","html","links","attachments"]],StructuralInfo["spam"],shuffle=True,random_state=42,test_size=0.15)

# Select the best hyper param alpha
alphalst=[0.1,1,10,100,1000,2000,5000,10000]
auclst=[]
accuracylst=[]
for a in alphalst:
    model=MultinomialNB(alpha=a)
    cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True)
    auclst.append(cross_result['test_roc_auc'].mean())
    accuracylst.append(cross_result['test_accuracy'].mean())
    
print(auclst)
print("The best alpha is",alphalst[np.argmax(auclst)])
print("Accuracy with best auc", accuracylst[np.argmax(auclst)])

print("The 5-fold auc of NB is",np.max(auclst))
print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])

model=MultinomialNB(alpha=np.argmax(auclst))
model.fit(x_train,y_train)
Y_pred=model.predict_proba(x_test)
print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))

# Subject model

import pickle
import sys 
sys.path.append("../../")
from E4525_ML import text

def subject_model(feature="set"):
    data_dir = "subject_feature"

    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    X_set=pickle.load(open( set_features_filename, "rb" ) )
    labels=pickle.load(open( labels_filename, "rb" ) )
    
    x_train,x_test,y_train,y_test=train_test_split(X_set,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    auclst=[]
    accuracylst=[]
    for a in alphalst:
        model=MultinomialNB(alpha=a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best alpha is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc of NB is",np.max(auclst))
    print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])
    
    model=MultinomialNB(alpha=alphalst[np.argmax(auclst)])
    model.fit(x_train,y_train)
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test
    
print("----------------------------------------------------------------")
subject_model(feature="set")
print("----------------------------------------------------------------")
subject_model(feature="count")
print("----------------------------------------------------------------")
subject_model(feature="tfidf")
print("----------------------------------------------------------------")

The best feature is tf-idf feature, which achieves 5-fold auc $0.983$ and accuracy $0.919$ on validation set and $0.927$ on test set

# Message body model

import numpy as np
import pandas as pd
from sklearn.naive_bayes import MultinomialNB,ComplementNB
from sklearn.model_selection import train_test_split, cross_validate

import pickle
import sys 
sys.path.append("../../")
from E4525_ML import text

def message_model(feature="set",n=-1,model_type="MB"):
    data_dir = "message_feature"

    set_features_filename=     data_dir+"/{}_features_{}.p".format(feature,-1)
    labels_filename=     data_dir+"/labels_{}.p".format(-1)

    X_set=pickle.load(open( set_features_filename, "rb" ) )
    labels=pickle.load(open( labels_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set=X_set[:n]
    
    x_train,x_test,y_train,y_test=train_test_split(X_set,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    auclst=[]
    accuracylst=[]
    for a in alphalst:
        if model_type=="MB":
            model=MultinomialNB(alpha=a)
        else: 
            model=ComplementNB(alpha=a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best alpha is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc of NB is",np.max(auclst))
    print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])
    
    if model_type=="MB":
        model=MultinomialNB(alpha=alphalst[np.argmax(auclst)])
    else: 
        model=ComplementNB(alpha=alphalst[np.argmax(auclst)])
    
    model.fit(x_train,y_train)
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test
    
print("----------------------------------------------------------------")
message_model(feature="set",n=-1)
print("----------------------------------------------------------------")
message_model(feature="count",n=-1)
print("----------------------------------------------------------------")
message_model(feature="tfidf",n=-1)
print("----------------------------------------------------------------")

# print("----------------------------------------------------------------")
# message_model(feature="set",n=-1,model_type="C")
# print("----------------------------------------------------------------")
# message_model(feature="count",n=-1,model_type="C")
# print("----------------------------------------------------------------")
# message_model(feature="tfidf",n=-1,model_type="C")
# print("----------------------------------------------------------------")

# Combined model (Combined features)

import numpy as np
import pandas as pd
from sklearn.naive_bayes import MultinomialNB,ComplementNB
from sklearn.model_selection import train_test_split, cross_validate
from sklearn.metrics import roc_auc_score

import pickle
import sys 
sys.path.append("../../")
from E4525_ML import text

from scipy.sparse import csr_matrix,hstack

def combined_model(feature="set",n=-1,model_type="MB"):
    
    # load message body features
    data_dir = "message_feature"

    set_features_filename=     data_dir+"/{}_features_{}.p".format(feature,-1)
    labels_filename=     data_dir+"/labels_{}.p".format(-1)

    X_set_body=pickle.load(open( set_features_filename, "rb" ) )
    labels=pickle.load(open( labels_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set_body=X_set_body[:n]
    
    # load subject features
    data_dir = "subject_feature"

    set_vectorizer_filename=   data_dir+"/{}_vectorizer.p".format(feature)
    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    # setVectorizer_subject=pickle.load(open( set_vectorizer_filename, "rb" ) )
    X_set_subject=pickle.load(open( set_features_filename, "rb" ) )
    
    if n!=-1:
        X_set_subject=X_set_subject[:n]
    
    # load structural info features
    StructuralInfo=pd.read_csv("email_05_structure.csv",index_col=[0])
    StructuralInfo=StructuralInfo.iloc[:,:-1]
    if n!=-1:
        StructuralInfo=StructuralInfo.values[:n]
    
    StructuralInfo=csr_matrix(StructuralInfo)
    
    # print(StructuralInfo.shape,X_set_subject.shape,X_set_body.shape)
    
    # Combine features together
    X_set=hstack([X_set_body,X_set_subject,StructuralInfo])
    
    # split train and test data set
    x_train,x_test,y_train,y_test=train_test_split(X_set,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    auclst=[]
    accuracylst=[]
    for a in alphalst:
        if model_type=="MB":
            model=MultinomialNB(alpha=a)
        else: 
            model=ComplementNB(alpha=a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best alpha is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc of NB is",np.max(auclst))
    print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])
    
    if model_type=="MB":
        model=MultinomialNB(alpha=alphalst[np.argmax(auclst)])
    else: 
        model=ComplementNB(alpha=alphalst[np.argmax(auclst)])
    
    model.fit(x_train,y_train)
    with open("./models/CombinedFeatNB_{}.p".format(feature),"wb") as f:
        pickle.dump(model,f)
    
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test

print("----------------------------------------------------------------")
combined_model(feature="set",n=-1)
print("----------------------------------------------------------------")
combined_model(feature="count",n=-1)
print("----------------------------------------------------------------")
combined_model(feature="tfidf",n=-1)
print("----------------------------------------------------------------")

# print("----------------------------------------------------------------")
# combined_model(feature="set",n=-1,model_type="C")
# print("----------------------------------------------------------------")
# combined_model(feature="count",n=-1,model_type="C")
# print("----------------------------------------------------------------")
# combined_model(feature="tfidf",n=-1,model_type="C")
# print("----------------------------------------------------------------")

# Combined model (Combined classifier)

import numpy as np
import pandas as pd
from sklearn.naive_bayes import MultinomialNB
from sklearn.model_selection import train_test_split, cross_validate
from sklearn.base import BaseEstimator
from sklearn.metrics import roc_auc_score

import pickle
import sys 
sys.path.append("../../")
from E4525_ML import text

from scipy.sparse import csr_matrix,hstack

class CombinedNB(BaseEstimator):
    def __init__(self,n1,n2,n3,a):
        self.n1=n1
        self.n2=n2
        self.n3=n3
        self.a=a
    
    def fit(self,X,y):
        self.model1=MultinomialNB(alpha=self.a)
        self.model1.fit(X[:,:self.n1],y)
        self.class_prior_=self.model1.class_count_/self.model1.class_count_.sum()
        
        self.model2=MultinomialNB(alpha=self.a)
        self.model2.fit(X[:,self.n1:(self.n1+self.n2)],y)
        self.model3=MultinomialNB(alpha=self.a)
        self.model3.fit(X[:,(self.n1+self.n2):(self.n1+self.n2+self.n3)],y)
        
    def predict_proba(self,X):
        p1=self.model1.predict_proba(X[:,:self.n1])
        p2=self.model2.predict_proba(X[:,self.n1:(self.n1+self.n2)])
        p3=self.model3.predict_proba(X[:,(self.n1+self.n2):(self.n1+self.n2+self.n3)])
        p=p1*p2*p3/(self.class_prior_)
        return p/(p.sum(axis=1)[:,np.newaxis])
    
    def predict(self,X):
        return np.argmax(self.predict_proba(X),axis=1)
    

def combined_model2(feature="set",n=-1):
    
    # load message body features
    data_dir = "message_feature"

    set_features_filename=     data_dir+"/{}_features_{}.p".format(feature,-1)
    labels_filename=     data_dir+"/labels_{}.p".format(-1)

    X_set_body=pickle.load(open( set_features_filename, "rb" ) )
    labels=pickle.load(open( labels_filename, "rb" ) )
    
    n_body=X_set_body.shape[1]
    
    if n!=-1:
        labels=labels[:n]
        X_set_body=X_set_body[:n]
    
    # load subject features
    data_dir = "subject_feature"

    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    X_set_subject=pickle.load(open( set_features_filename, "rb" ) )
    
    if n!=-1:
        X_set_subject=X_set_subject[:n]
        
    n_sub=X_set_subject.shape[1]
    
    # load structural info features
    StructuralInfo=pd.read_csv("email_05_structure.csv",index_col=[0])
    StructuralInfo=StructuralInfo.iloc[:,:-1]
    if n!=-1:
        StructuralInfo=StructuralInfo.values[:n]
    
    StructuralInfo=csr_matrix(StructuralInfo)
    
    n_structural=StructuralInfo.shape[1]
    
    print(n_body,n_sub,n_structural)
    
    # Combine features together
    X_set=hstack([X_set_body,X_set_subject,StructuralInfo])
    
    # split train and test data set
    x_train,x_test,y_train,y_test=train_test_split(X_set,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    auclst=[]
    accuracylst=[]
    for a in alphalst:
        model=CombinedNB(n_body,n_sub,n_structural,a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best alpha is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc of NB is",np.max(auclst))
    print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])
    
    model=CombinedNB(n_body,n_sub,n_structural,a=alphalst[np.argmax(auclst)])
    model.fit(x_train,y_train)
    
    with open("./models/CombinedProbNB_{}.p".format(feature),"wb") as f:
        pickle.dump(model,f)
        
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test

print("----------------------------------------------------------------")
combined_model2(feature="set",n=-1)
print("----------------------------------------------------------------")
combined_model2(feature="count",n=-1)
print("----------------------------------------------------------------")
combined_model2(feature="tfidf",n=-1)
