==> 3.0_other_stats_learning_models.ipynb <==
# SVM

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split, cross_validate
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import GridSearchCV

import pickle
import sys 
sys.path.append("../../")
from E4525_ML import text

from sklearn.svm import SVC

def SVM_subject_model(feature="set",n=-1):
    
    # load subject features
    data_dir = "subject_feature"

    set_vectorizer_filename=   data_dir+"/{}_vectorizer.p".format(feature)
    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    labels=pickle.load(open( labels_filename, "rb" ) )
    X_set_subject=pickle.load(open( set_features_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set_subject=X_set_subject[:n]
    
    # split train and test data set
    x_train,x_test,y_train,y_test=train_test_split(X_set_subject,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    param_grid={"C":[1],"gamma":[0.1,0.5,1,10,20]}
    model=SVC(probability=True)
    g=GridSearchCV(model,param_grid,cv=5, scoring=["roc_auc","accuracy"],n_jobs=-1,refit="roc_auc")
    g.fit(x_train, y_train)
    
    print("{} feature: ".format(feature))
    print(g.cv_results_['mean_test_roc_auc'])
    print("The best alpha is",g.best_estimator_.get_params()["C"])
    print("The best gamma is",g.best_estimator_.get_params()["gamma"])
    print("The 5-fold auc is",g.cv_results_['mean_test_roc_auc'].max())
    print("The 5-fold acc is",g.cv_results_['mean_test_accuracy'].max())
    
    Y_pred=g.best_estimator_.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(g.best_estimator_.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test

print("----------------------------------------------------------------")
SVM_subject_model(feature="set",n=10000)
print("----------------------------------------------------------------")
SVM_subject_model(feature="count",n=10000)
print("----------------------------------------------------------------")
SVM_subject_model(feature="tfidf",n=10000)

# NB model

from sklearn.naive_bayes import MultinomialNB

def subject_model(feature="set",n=-1):
    data_dir = "subject_feature"

    set_vectorizer_filename=   data_dir+"/{}_vectorizer.p".format(feature)
    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    X_set=pickle.load(open( set_features_filename, "rb" ) )
    labels=pickle.load(open( labels_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set=X_set[:n]
    
    x_train,x_test,y_train,y_test=train_test_split(X_set,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    auclst=[]
    accuracylst=[]
    for a in alphalst:
        model=MultinomialNB(alpha=a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True,n_jobs=-1)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best alpha is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc of NB is",np.max(auclst))
    print("The 5-fold accuracy of NB is",accuracylst[np.argmax(auclst)])
    
    model=MultinomialNB(alpha=alphalst[np.argmax(auclst)])
    model.fit(x_train,y_train)
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test
    
print("----------------------------------------------------------------")
subject_model(feature="set",n=10000)
print("----------------------------------------------------------------")
subject_model(feature="count",n=10000)
print("----------------------------------------------------------------")
subject_model(feature="tfidf",n=10000)

# Logistic

from sklearn.linear_model import LogisticRegression

def Logistic_subject_model(feature="set",n=-1):
    
    # load subject features
    data_dir = "subject_feature"

    set_vectorizer_filename=   data_dir+"/{}_vectorizer.p".format(feature)
    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    labels=pickle.load(open( labels_filename, "rb" ) )
    X_set_subject=pickle.load(open( set_features_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set_subject=X_set_subject[:n]
    
    # split train and test data set
    x_train,x_test,y_train,y_test=train_test_split(X_set_subject,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    alphalst=[1e-4,1e-3,0.1,1,10,100,1000]
    gammalst=[]
    auclst=[]
    accuracylst=[]
    modellst=[]
    for a in alphalst:
        model=LogisticRegression(C=a)
        cross_result=cross_validate(model,x_train,y_train,scoring=["roc_auc","accuracy"],cv=5,return_train_score=True,n_jobs=-1)
        auclst.append(cross_result['test_roc_auc'].mean())
        accuracylst.append(cross_result['test_accuracy'].mean())
        modellst.append(model)

    print("{} feature: ".format(feature))
    print(auclst)
    print("The best C is",alphalst[np.argmax(auclst)])
    
    print("The 5-fold auc is",np.max(auclst))
    print("The 5-fold accuracy is",accuracylst[np.argmax(auclst)])
    
    model=LogisticRegression(C=np.max(auclst))
    model.fit(x_train,y_train)
    Y_pred=model.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(model.predict(x_test)==y_test))
    
    return x_train,x_test,y_train,y_test

print("----------------------------------------------------------------")
SVM_subject_model(feature="set",n=10000)
print("----------------------------------------------------------------")
SVM_subject_model(feature="count",n=10000)
print("----------------------------------------------------------------")
SVM_subject_model(feature="tfidf",n=10000)

# KNN

from sklearn.neighbors import KNeighborsClassifier

def KNN_subject_model(feature="set",n=-1):
    
    # load subject features
    data_dir = "subject_feature"

    set_vectorizer_filename=   data_dir+"/{}_vectorizer.p".format(feature)
    set_features_filename=     data_dir+"/{}_features.p".format(feature)
    labels_filename=     data_dir+"/labels.p"

    labels=pickle.load(open( labels_filename, "rb" ) )
    X_set_subject=pickle.load(open( set_features_filename, "rb" ) )
    
    if n!=-1:
        labels=labels[:n]
        X_set_subject=X_set_subject[:n]
    
    # split train and test data set
    x_train,x_test,y_train,y_test=train_test_split(X_set_subject,labels,shuffle=True,random_state=42,test_size=0.15)
    
    # Select the best hyper param alpha
    param_grid={"n_neighbors":[1,3,5,7,10,20,50,100] }
    model=KNeighborsClassifier()
    g=GridSearchCV(model,param_grid,cv=5, scoring=["roc_auc","accuracy"],n_jobs=-1,refit="roc_auc")
    g.fit(x_train, y_train)
    
    print("{} feature: ".format(feature))
    print(g.cv_results_['mean_test_roc_auc'])
    print("The best n_neighbors is",g.best_estimator_.get_params()["n_neighbors"])
    print("The 5-fold auc is",g.cv_results_['mean_test_roc_auc'].max())
    print("The 5-fold acc is",g.cv_results_['mean_test_accuracy'].max())
    
    Y_pred=g.best_estimator_.predict_proba(x_test)
    print("auc on test set is",roc_auc_score(y_test,Y_pred[:,1]))
    print("Accuracy on test set is",np.mean(g.best_estimator_.predict(x_test)==y_test))

    return x_train,x_test,y_train,y_test

print("----------------------------------------------------------------")
KNN_subject_model(feature="set",n=10000)
print("----------------------------------------------------------------")
KNN_subject_model(feature="count",n=10000)
print("----------------------------------------------------------------")
KNN_subject_model(feature="tfidf",n=10000)
